@@ -19,16 +19,23 @@
     "@fortawesome/fontawesome-svg-core": "^6.5.1",
     "@fortawesome/free-regular-svg-icons": "^6.5.1",
     "@fortawesome/free-solid-svg-icons": "^6.5.1",
-    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@fortawesome/react-fontawesome": "^3.1.1",
     "@module-federation/vite": "^1.0.10",
-    "@rjsf/bootstrap-4": "^5.24.13",
+    "@reduxjs/toolkit": "^2.3.0",
     "@rjsf/core": "^5.24.13",
     "@rjsf/utils": "^5.24.13",
     "@rjsf/validator-ajv8": "^5.24.13",
     "@signalk/server-admin-ui-dependencies": "1.0.1",
+    "@types/escape-html": "^1.0.4",
+    "@types/lodash.remove": "^4.7.9",
+    "@types/lodash.set": "^4.3.9",
+    "@types/lodash.uniq": "^4.5.9",
+    "@types/react": "^19.0.0",
+    "@types/react-dom": "^19.0.0",
+    "@types/react-redux": "^7.1.34",
     "@vitejs/plugin-react": "^4.3.4",
     "ansi-to-html": "^0.6.14",
-    "bootstrap": "^4.5.3",
+    "bootstrap": "^5.3.3",
     "buffer": "^6.0.3",
     "escape-html": "^1.0.3",
     "font-awesome": "^4.7.0",
@@ -37,23 +44,20 @@
     "lodash.set": "^4.3.2",
     "lodash.uniq": "^4.5.0",
     "moment": "^2.29.1",
-    "react": "^16.14.0",
-    "react-bootstrap": "^1.6.8",
-    "react-copy-to-clipboard": "^5.0.3",
-    "react-dom": "^16.14.0",
-    "react-html-parser": "^2.0.2",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "html-react-parser": "^5.2.5",
     "react-infinite-scroll-component": "^6.1.0",
     "react-json-tree": "^0.20.0",
-    "react-redux": "^5.1.2",
-    "react-router-dom": "^4.3.1",
-    "react-select": "^3.1.0",
-    "reactstrap": "^5.0.0",
+    "react-redux": "^9.1.2",
+    "react-router-dom": "^6.28.0",
+    "react-select": "^5.10.2",
+    "reactstrap": "^9.2.3",
     "reconnecting-websocket": "^4.4.0",
-    "redux": "^3.7.2",
-    "redux-thunk": "2.3.0",
     "rollup-plugin-visualizer": "^5.12.0",
     "sass": "^1.81.0",
     "simple-line-icons": "^2.5.5",
+    "typescript": "^5.7.2",
     "vite": "^6.0.5"
   },
   "scripts": {
