@@ -25,6 +25,7 @@ class AccessRequests extends Component {
       accessRequestsApproving: [],
       accessRequestsDenying: []
     }
+    this.selectedRequestRef = React.createRef()
     this.handleRequestChange = this.handleRequestChange.bind(this)
   }
 
@@ -73,7 +74,7 @@ class AccessRequests extends Component {
         selectedIndex: index
       },
       () => {
-        this.refs['selectedRequest'].scrollIntoView()
+        this.selectedRequestRef.current?.scrollIntoView()
       }
     )
   }
@@ -147,7 +148,7 @@ class AccessRequests extends Component {
             </Card>
 
             {this.state.selectedRequest && (
-              <div ref="selectedRequest">
+              <div ref={this.selectedRequestRef}>
                 <Card>
                   <CardHeader>
                     <i className="fa fa-align-justify"></i>Request
@@ -242,10 +243,10 @@ class AccessRequests extends Component {
                   <CardFooter>
                     <Row
                       className={
-                        'ml-0 mr-0 d-flex justify-content-between justify-content-sm-start'
+                        'ms-0 me-0 d-flex justify-content-between justify-content-sm-start'
                       }
                     >
-                      <Col xs="4" md="4" lg="2" className={'pl-0 pr-0 pr-md-2'}>
+                      <Col xs="4" md="4" lg="2" className={'ps-0 pe-0 pe-md-2'}>
                         <Button
                           size="md"
                           color="success"
@@ -271,12 +272,12 @@ class AccessRequests extends Component {
                         xs="4"
                         md="8"
                         lg="3"
-                        className={'pl-2 pl-lg-1 pr-0 pr-md-2'}
+                        className={'ps-2 ps-lg-1 pe-0 pe-md-2'}
                       >
                         <Button
                           size="md"
                           color="danger"
-                          className="float-right float-sm-left"
+                          className="float-end float-sm-start"
                           onClick={this.handleAccessRequest.bind(
                             this,
                             this.state.selectedRequest.accessIdentifier,
