@@ -9,7 +9,6 @@ import {
   Button,
   Input,
   InputGroup,
-  InputGroupAddon,
   InputGroupText,
   Label
 } from 'reactstrap'
@@ -146,11 +145,9 @@ class Login extends Component {
                         <h1>Login</h1>
                         <p className="text-muted">Sign In to your account</p>
                         <InputGroup className="mb-3">
-                          <InputGroupAddon addonType="prepend">
-                            <InputGroupText>
-                              <i className="icon-user" />
-                            </InputGroupText>
-                          </InputGroupAddon>
+                          <InputGroupText>
+                            <i className="icon-user" />
+                          </InputGroupText>
                           <Input
                             type="text"
                             name="username"
@@ -160,11 +157,9 @@ class Login extends Component {
                           />
                         </InputGroup>
                         <InputGroup className="mb-4">
-                          <InputGroupAddon addonType="prepend">
-                            <InputGroupText>
-                              <i className="icon-lock" />
-                            </InputGroupText>
-                          </InputGroupAddon>
+                          <InputGroupText>
+                            <i className="icon-lock" />
+                          </InputGroupText>
                           <Input
                             type="password"
                             name="password"
@@ -211,7 +206,7 @@ class Login extends Component {
                               Login
                             </Button>
                           </Col>
-                          <Col xs="6" className="text-right">
+                          <Col xs="6" className="text-end">
                             {this.state.loginErrorMessage && (
                               <p className="text-danger">
                                 {this.state.loginErrorMessage}
