@@ -10,7 +10,6 @@ import {
   Form,
   Input,
   InputGroup,
-  InputGroupAddon,
   InputGroupText
 } from 'reactstrap'
 
@@ -98,9 +97,7 @@ class Register extends Component {
                     <Form>
                       <p className="text-muted">Create your account</p>
                       <InputGroup className="mb-3">
-                        <InputGroupAddon addonType="prepend">
-                          <InputGroupText>@</InputGroupText>
-                        </InputGroupAddon>
+                        <InputGroupText>@</InputGroupText>
                         <Input
                           name="email"
                           type="text"
@@ -109,11 +106,9 @@ class Register extends Component {
                         />
                       </InputGroup>
                       <InputGroup className="mb-3">
-                        <InputGroupAddon addonType="prepend">
-                          <InputGroupText>
-                            <i className="icon-lock" />
-                          </InputGroupText>
-                        </InputGroupAddon>
+                        <InputGroupText>
+                          <i className="icon-lock" />
+                        </InputGroupText>
                         <Input
                           name="password"
                           type="password"
@@ -122,11 +117,9 @@ class Register extends Component {
                         />
                       </InputGroup>
                       <InputGroup className="mb-0">
-                        <InputGroupAddon addonType="prepend">
-                          <InputGroupText>
-                            <i className="icon-lock" />
-                          </InputGroupText>
-                        </InputGroupAddon>
+                        <InputGroupText>
+                          <i className="icon-lock" />
+                        </InputGroupText>
                         <Input
                           name="confirmPassword"
                           type="password"
